Make m_bunchCrossingTool of src/SCTHitEffMonTool.cxx public tool again. (SCT_Monitoring-00-06-98)
	* Make m_bunchCrossingTool of src/SCTHitEffMonTool.cxx public tool again.
	* tag as SCT_Monitoring-00-06-98

2016-12-03 Susumu Oda <Susumu.Oda@cern.ch>
	* Revert TProfile2D_LW to TProfile2D in SCTHitsNoiseMonTool.
	* tag as SCT_Monitoring-00-06-97

2016-11-29 Shunsuke Honda <Shunsuke.Honda@cern.ch>
	* Make NBINS_LBs common variable in the SCT_Monitoring in SCT_MonitoringNumbers
	* Change NBINS_LBs 2000 to 3000
	* tag as SCT_Monitoring-00-06-96

2016-11-28 Shunsuke Honda <Shunsuke.Honda@cern.ch>
	* Change variables TH1D to TH1F and TH? to TH?_LW in the SCTHitsNoiseMonTool(ATLASRECTS-3440)
	* tag as SCT_Monitoring-00-06-95

2016-11-28 Shunsuke Honda <Shunsuke.Honda@cern.ch>
	* Change variables double to float
	* tag as SCT_Monitoring-00-06-94

...
(Long ChangeLog diff - truncated)


Former-commit-id: e60685f668bf89456b4469fc60de94c7eb73b69b

+++ InnerDetector/InDetMonitoring/SCT_Monitoring/SCT_Monitoring/SCTTracksMonTool.h
@@ -178,7 +178,7 @@ private:
   //@name Service methods
   //@{
   /// Calculate Pull value for MeasuredAtPlane TrackStates
-  double calculatePull(const double, const double, const double);
+  float calculatePull(const float, const float, const float);
   
   ///Factory + register for the 2D histos, returns whether successfully registered
   StatusCode h2Factory(const std::string & name, const std::string & title, const SCT_Monitoring::Bec bec, MonGroup & registry, VecH2_t & storageVector);
